pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "deep-earth-degasation"
version = "0.1.0"
description = "Candidate detection MVP for surface manifestations compatible with possible deep degassing"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "Deep Earth Degasation project" }]
dependencies = [
  "geopandas>=0.14",
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.2",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "rasterio>=1.3",
  "rioxarray>=0.15",
  "scikit-image>=0.22",
  "scipy>=1.12",
  "shapely>=2.0",
  "typer>=0.12",
  "xarray>=2024.1.0",
]

[project.optional-dependencies]
dev = [
  "basedpyright>=1.25",
  "pytest>=8.0",
  "ruff>=0.8",
]
ml = [
  "catboost>=1.2",
  "lightgbm>=4.3",
  "scikit-learn>=1.4",
  "shap>=0.45",
]

[project.scripts]
degasation = "deep_earth_degasation.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/deep_earth_degasation"]

[tool.ruff]
line-length = 100
src = ["src"]
target-version = "py311"

[tool.ruff.lint]
select = [
  "A",
  "B",
  "E",
  "F",
  "I",
  "N",
  "PT",
  "RUF",
  "UP",
  "W",
]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.basedpyright]
include = ["src", "tests"]
pythonVersion = "3.11"
typeCheckingMode = "standard"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportAny = false
reportExplicitAny = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
